+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Adds missing typescript definitions for `Utility` class
 - Fixs incorrect typescript definitions for `Address`, `EndShipper`m and `Webhook` classes methods
+- Moves `Utils` to EasyPostClient class
 
 ## v6.4.0 (2023-05-02)